+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGES
 
 **2.0.5**
-
+* Add symbolPadding, symbolHeight and symbolWidth in `ChartLegend`
 
 **2.0.4**
 * Add margin, padding, itemMarginTop, and itemMarginBottom in `ChartLegend`